Add message-id & message body validation & merkle tree metadata support to the Hyperlane message plugin.
