+++ cabal.project
@@ -75,8 +75,8 @@ package yet-another-logger
 source-repository-package
     type: git
     location: https://github.com/kadena-io/pact.git
-    tag: 9ccef1fbeff545f07864896094eb2e2bd4c5ffb3
-    --sha256: 0pai52mvyikhm527if4gxa1nwsmva3wg3nhvida7qg48rchjrxry
+    tag: f1d39bf2a5b3fcc932f5bb122784a7b6fbe35fae
+    --sha256: sha256-VtYgCMyj6T04ZCqiWYvKPkTW1nZQf5j4Ci1ccZy0ETo=
 
 source-repository-package
     type: git
